Move database utils to separate file
